requirements.txt
# League of Legends API Client Requirements
# Generated from conda league environment
# Install with: pip install -r requirements.txt

# Core dependencies
requests==2.32.4
python-dotenv==1.1.1
urllib3==2.5.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10

# Development and testing tools
black==25.1.0
flake8==7.3.0
mypy==1.17.0
pytest==8.4.1

# Supporting packages
click==8.2.1
iniconfig==2.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
setuptools==78.1.1
typing_extensions==4.14.1
wheel==0.45.1